fix merged in code to work with DB refactor
