+++ src/main/java/edu/wpi/aquamarine_axolotls/pathplanning/DepthFirstSearch.java
@@ -44,7 +44,7 @@ public class DepthFirstSearch extends AbsAlgorithmMethod {
                         edgeMap.get(j).get("ENDNODE")
                 ));
             }
-        } catch (SQLException | IOException | URISyntaxException e) {
+        } catch (SQLException | IOException e) {
             e.printStackTrace();
         }
     }
