+++ src/main/java/edu/wpi/aquamarine_axolotls/pathplanning/SearchAlgorithm.java
@@ -49,7 +49,7 @@ public class SearchAlgorithm{
                         edgeMap.get(j).get("ENDNODE")
                 ));
             }
-        } catch (SQLException | IOException | URISyntaxException e) {
+        } catch (SQLException | IOException e) {
             e.printStackTrace();
         }
     }
