+++ src/main/java/edu/wpi/aquamarine_axolotls/views/mapping/GenericMap.java
@@ -62,7 +62,7 @@ public class GenericMap extends GenericPage {
      */
     public void startUp(){
         try {
-            db = new DatabaseController();
+            db = DatabaseController.getInstance();
 
   /*
             }*/
@@ -89,7 +89,7 @@ public class GenericMap extends GenericPage {
             //drawFloor(FLOOR);
             drawNodesAndFloor ( FLOOR, Color.BLUE );
             zoom = 1;
-        } catch (SQLException | IOException | URISyntaxException e) {
+        } catch (SQLException | IOException e) {
             e.printStackTrace();
         }
     }
