+++ src/main/java/edu/wpi/aquamarine_axolotls/views/servicerequests/LanguageInterpreter.java
@@ -2,6 +2,7 @@ package edu.wpi.aquamarine_axolotls.views.servicerequests;
 
 import com.jfoenix.controls.*;
 import edu.wpi.aquamarine_axolotls.db.*;
+import edu.wpi.aquamarine_axolotls.db.enums.SERVICEREQUEST;
 import edu.wpi.aquamarine_axolotls.views.servicerequests.GenericServiceRequest;
 import javafx.collections.FXCollections;
 import javafx.event.ActionEvent;
